fix:abstract test helpers
